♻️ Simplify implementation with --force option

+++ src/function/ensure_file_template.sh
@@ -1,12 +1,42 @@
 #!/usr/bin/env bash
 
+# Ensure file is created using template
+#
+# Usage :
+#   laptop::ensure_file_template <template> <target>
+# Options:
+#   --force  Force to overwrite file if present
+#
 laptop::ensure_file_template() {
   local template="$1"
   local target="$2"
+  local force=0
+  local positional_args=()
 
-  laptop::step_start "- Ensure file '$target'"
-  laptop::step_eval "\
-  mkdir -p $(quote "$(dirname "$target")") && \
-  cp -f $(quote "$template") $(quote "$target") \
-  "
+  while [[ $# -gt 0 ]]; do
+    case $1 in
+      -f|--force)
+        force=1
+        # shift
+        ;;
+      -*)
+        echo "Unknown option $1"
+        exit 1
+        ;;
+      *)
+        positional_args+=("$1") # save positional arg
+        shift # past argument
+        ;;
+    esac
+  done
+
+  laptop::step_start "- Ensure file '$target'$([ "$force" -eq 1 ] && echo ' (force)') "
+  if [ "$force" -eq 0 ] && [ -f "$target" ]; then
+    laptop::step_pass
+  else
+    laptop::step_eval "\
+    mkdir -p $(quote "$(dirname "$target")") && \
+    cp -f $(quote "$template") $(quote "$target") \
+    "
+  fi
 }
